+++ base.py
@@ -1,14 +1,6 @@
 from pyfirmata2 import Arduino, util
 import serial.tools.list_ports
 
-# def find_arduino(name):
-#     """Détecte automatiquement le port de l'Arduino et retourne son nom de port."""
-#     ports = serial.tools.list_ports.comports()
-#     for port in ports:
-#         if name in port.description or "tty" in port.device or "COM" in port.device:
-#             return port
-#     raise IOError("Aucun Arduino détecté. Vérifiez la connexion.")
-
 def connect_to_arduino(name):
     """Établit la connexion à l'Arduino."""
     try:
